+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - status qr code widget - show pin in monospace
 - bridges widget - show bridges as restarting on full restart
 - do not hide search button when no search results found
+- close settings restart toast if any restart button clicked
 
 ### Other Changes